Add LAPPDLoadTXT to load ASCII data with other LAPPD Tools. Add LAPPDASCIIAna for ASCII analysis with new tools. Most of the config files were copied from LAPPDProcessedAna
 Changes to be committed:
	modified:   UserTools/Factory/Factory.cpp
	new file:   UserTools/LAPPDLoadTXT/LAPPDLoadTXT.cpp
	new file:   UserTools/LAPPDLoadTXT/LAPPDLoadTXT.h
	new file:   UserTools/LAPPDLoadTXT/README.md
	modified:   UserTools/LAPPDPlots/LAPPDPlots.cpp
	modified:   UserTools/Unity.h
	new file:   configfiles/LAPPDASCIIAna/ConfigGeo
	new file:   configfiles/LAPPDASCIIAna/ConfigPlot
	new file:   configfiles/LAPPDASCIIAna/ConfigPreProcess
	new file:   configfiles/LAPPDASCIIAna/ConfigStoreReadIn
	new file:   configfiles/LAPPDASCIIAna/ConfigTXT
	new file:   configfiles/LAPPDASCIIAna/Configs
	new file:   configfiles/LAPPDASCIIAna/FindMrdTracksConfig
	new file:   configfiles/LAPPDASCIIAna/LoadANNIEEventConfig
	new file:   configfiles/LAPPDASCIIAna/ProcessedLAPPDFilterConfig
	new file:   configfiles/LAPPDASCIIAna/TimeClusteringConfig
	new file:   configfiles/LAPPDASCIIAna/ToolChainConfig
	new file:   configfiles/LAPPDASCIIAna/ToolsConfig
	new file:   configfiles/LAPPDASCIIAna/knownlist

+++ UserTools/LAPPDPlots/LAPPDPlots.cpp
@@ -296,7 +296,9 @@ bool LAPPDPlots::Execute()
       if (!drawTriggerChannel)
         nstrips = 28;
       const double BinHistMinConst = BinHistMin;
-      const double BinHistMaxConst = BinHistMax;
+      //cout<<"min is "<<BinHistMin<<", "<<BinHistMinConst<<endl;
+      const int BinHistMaxConst = BinHistMax;
+      //cout<<"max is "<<BinHistMax<<", "<<BinHistMaxConst<<endl;
       const int BinHistNumberConst = BinHistNumber;
       TH2D *h = new TH2D(HistoName, HistoName, BinHistNumberConst, BinHistMinConst, BinHistMaxConst, nstrips, 0 - 0.5, nstrips - 0.5);
       if (LAPPDPlotsVerbosity > 3)
